refactor(fe): move isOpenIdCancelError tests; close sidebar on any nav
- Relocate the six isOpenIdCancelError test cases from AuthWizard.test.ts
  into openID.test.ts (where the function lives) and delete the
  AuthWizard test file, which never imported AuthWizard.
- Drop the pathname-comparison guard in /manage's afterNavigate; the
  mobile sidebar should close on any navigation, since users can't
  interact with the page while it's open.
